+++ connecteur/fakeIparapheur/FakeIparapheur.class.php
@@ -181,7 +181,7 @@ class FakeIparapheur extends SignatureConnecteur
      * @param $signature
      * @return Fichier
      */
-    public function getBordereauFromSignature($signature): Fichier
+    public function getBordereauFromSignature($signature): ?Fichier
     {
         $file = new Fichier();
         $file->filename = $signature['nom_document'];